feat: implement get all transaction (owner / cashier) including filtering option

+++ DTOs/TransactionDto/TransactionDetailListDto.cs
@@ -5,8 +5,8 @@
         public Guid ProductId { get; set; }
         public string ProductName { get; set; }
         public int Quantity { get; set; }
-        public long Price { get; set; }
-        public long Subtotal => Quantity * Price;
+        public decimal Price { get; set; }
+        public decimal Subtotal => Quantity * Price;
     }
 
 }
